storage: Fix iscsi-direct volume size for volumes > 4GiB
Both block_size and nb_block are unit32_t.

Message-Id: <13a14a69663e89354628222596635f3d566b9bd3.1551363720.git.jdenemar@redhat.com>

+++ src/storage/storage_backend_iscsi_direct.c
@@ -321,8 +321,8 @@ virISCSIDirectRefreshVol(virStoragePoolObjPtr pool,
     if (virISCSIDirectGetVolumeCapacity(iscsi, lun, &block_size, &nb_block) < 0)
         return -1;
 
-    vol->target.capacity = block_size * nb_block;
-    vol->target.allocation = block_size * nb_block;
+    vol->target.capacity = (unsigned long long)block_size * nb_block;
+    vol->target.allocation = (unsigned long long)block_size * nb_block;
     def->capacity += vol->target.capacity;
     def->allocation += vol->target.allocation;
